fix: added contract information in the dpp
